requirements.txt
inquirer==3.2.4
python-dotenv==1.0.1
Requests==2.31.0
psycopg2-binary==2.9.9